鸡哥智能协同云图库项目 ---引入 Sa-Token 权限校验框架

+++ src/main/java/com/jige/jigepicturebackend/controller/SpaceUserController.java
@@ -2,12 +2,15 @@ package com.jige.jigepicturebackend.controller;
 
 import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.util.ObjUtil;
+import cn.hutool.core.util.ObjectUtil;
 import com.jige.jigepicturebackend.common.BaseResponse;
 import com.jige.jigepicturebackend.common.DeleteRequest;
 import com.jige.jigepicturebackend.common.ResultUtils;
 import com.jige.jigepicturebackend.exception.BusinessException;
 import com.jige.jigepicturebackend.exception.ErrorCode;
 import com.jige.jigepicturebackend.exception.ThrowUtils;
+import com.jige.jigepicturebackend.manager.auth.annotation.SaSpaceCheckPermission;
+import com.jige.jigepicturebackend.manager.auth.model.SpaceUserPermissionConstant;
 import com.jige.jigepicturebackend.model.dto.spaceuser.SpaceUserAddRequest;
 import com.jige.jigepicturebackend.model.dto.spaceuser.SpaceUserEditRequest;
 import com.jige.jigepicturebackend.model.dto.spaceuser.SpaceUserQueryRequest;
@@ -27,44 +30,46 @@ import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
+/**
+ * 空间成员管理
+ */
 @RestController
 @RequestMapping("/spaceUser")
 @Slf4j
 public class SpaceUserController {
-    @Resource
-    private UserService userService;
 
     @Resource
     private SpaceUserService spaceUserService;
 
+    @Resource
+    private UserService userService;
+
     /**
      * 添加成员到空间
-     *
-     * @param spaceUserAddRequest
-     * @return
      */
     @PostMapping("/add")
+    @SaSpaceCheckPermission(value = SpaceUserPermissionConstant.SPACE_USER_MANAGE)
     public BaseResponse<Long> addSpaceUser(@RequestBody SpaceUserAddRequest spaceUserAddRequest, HttpServletRequest request) {
         ThrowUtils.throwIf(spaceUserAddRequest == null, ErrorCode.PARAMS_ERROR);
-        long spaceUserId = spaceUserService.addSpaceUser(spaceUserAddRequest);
-        return ResultUtils.success(spaceUserId);
+        long id = spaceUserService.addSpaceUser(spaceUserAddRequest);
+        return ResultUtils.success(id);
     }
 
     /**
      * 从空间移除成员
-     *
-     * @param deleteRequest
-     * @param request
-     * @return
      */
     @PostMapping("/delete")
-    public BaseResponse<Boolean> deleteSpaceUser(@RequestBody DeleteRequest deleteRequest, HttpServletRequest request) {
-        ThrowUtils.throwIf(deleteRequest == null || deleteRequest.getId() <= 0, ErrorCode.PARAMS_ERROR);
-        Long id = deleteRequest.getId();
-        //判断是否存在
-        SpaceUser spaceUser = spaceUserService.getById(id);
-        ThrowUtils.throwIf(spaceUser == null, ErrorCode.NOT_FOUND_ERROR);
-        //操作数据库
+    @SaSpaceCheckPermission(value = SpaceUserPermissionConstant.SPACE_USER_MANAGE)
+    public BaseResponse<Boolean> deleteSpaceUser(@RequestBody DeleteRequest deleteRequest,
+                                                 HttpServletRequest request) {
+        if (deleteRequest == null || deleteRequest.getId() <= 0) {
+            throw new BusinessException(ErrorCode.PARAMS_ERROR);
+        }
+        long id = deleteRequest.getId();
+        // 判断是否存在
+        SpaceUser oldSpaceUser = spaceUserService.getById(id);
+        ThrowUtils.throwIf(oldSpaceUser == null, ErrorCode.NOT_FOUND_ERROR);
+        // 操作数据库
         boolean result = spaceUserService.removeById(id);
         ThrowUtils.throwIf(!result, ErrorCode.OPERATION_ERROR);
         return ResultUtils.success(true);
@@ -72,79 +77,71 @@ public class SpaceUserController {
 
     /**
      * 查询某个成员在某个空间的信息
-     *
-     * @param spaceUserQueryRequest
-     * @param request
-     * @return
      */
     @PostMapping("/get")
+    @SaSpaceCheckPermission(value = SpaceUserPermissionConstant.SPACE_USER_MANAGE)
     public BaseResponse<SpaceUser> getSpaceUser(@RequestBody SpaceUserQueryRequest spaceUserQueryRequest) {
-        //参数校验
+        // 参数校验
         ThrowUtils.throwIf(spaceUserQueryRequest == null, ErrorCode.PARAMS_ERROR);
-        Long userId = spaceUserQueryRequest.getUserId();
         Long spaceId = spaceUserQueryRequest.getSpaceId();
-        ThrowUtils.throwIf(ObjUtil.hasEmpty(userId, spaceId), ErrorCode.PARAMS_ERROR);
-        //查询数据库
+        Long userId = spaceUserQueryRequest.getUserId();
+        ThrowUtils.throwIf(ObjectUtil.hasEmpty(spaceId, userId), ErrorCode.PARAMS_ERROR);
+        // 查询数据库
         SpaceUser spaceUser = spaceUserService.getOne(spaceUserService.getQueryWrapper(spaceUserQueryRequest));
         ThrowUtils.throwIf(spaceUser == null, ErrorCode.NOT_FOUND_ERROR);
         return ResultUtils.success(spaceUser);
     }
 
-
     /**
      * 查询成员信息列表
-     *
-     * @param spaceUserQueryRequest
-     * @param request
-     * @return
      */
     @PostMapping("/list")
-    public BaseResponse<List<SpaceUserVO>> listSpaceUserVo(@RequestBody SpaceUserQueryRequest spaceUserQueryRequest, HttpServletRequest request) {
+    @SaSpaceCheckPermission(value = SpaceUserPermissionConstant.SPACE_USER_MANAGE)
+    public BaseResponse<List<SpaceUserVO>> listSpaceUser(@RequestBody SpaceUserQueryRequest spaceUserQueryRequest,
+                                                         HttpServletRequest request) {
         ThrowUtils.throwIf(spaceUserQueryRequest == null, ErrorCode.PARAMS_ERROR);
-        List<SpaceUser> spaceUserList = spaceUserService.list(spaceUserService.getQueryWrapper(spaceUserQueryRequest));
-        ThrowUtils.throwIf(CollUtil.isEmpty(spaceUserList), ErrorCode.NOT_FOUND_ERROR);
+        List<SpaceUser> spaceUserList = spaceUserService.list(
+                spaceUserService.getQueryWrapper(spaceUserQueryRequest)
+        );
         return ResultUtils.success(spaceUserService.getSpaceUserVOList(spaceUserList));
     }
 
     /**
      * 编辑成员信息（设置权限）
-     *
-     * @param spaceUserEditRequest
-     * @return
      */
     @PostMapping("/edit")
-    public BaseResponse<Boolean> editSpaceUser(@RequestBody SpaceUserEditRequest spaceUserEditRequest, HttpServletRequest request) {
+    @SaSpaceCheckPermission(value = SpaceUserPermissionConstant.SPACE_USER_MANAGE)
+    public BaseResponse<Boolean> editSpaceUser(@RequestBody SpaceUserEditRequest spaceUserEditRequest,
+                                               HttpServletRequest request) {
         if (spaceUserEditRequest == null || spaceUserEditRequest.getId() <= 0) {
             throw new BusinessException(ErrorCode.PARAMS_ERROR);
         }
-        //将实体类和 DTO 转换
+        // 将实体类和 DTO 进行转换
         SpaceUser spaceUser = new SpaceUser();
         BeanUtils.copyProperties(spaceUserEditRequest, spaceUser);
-        //数据校验
+        // 数据校验
         spaceUserService.validSpaceUser(spaceUser, false);
-        //判断是否存在
-        Long id = spaceUserEditRequest.getId();
+        // 判断是否存在
+        long id = spaceUserEditRequest.getId();
         SpaceUser oldSpaceUser = spaceUserService.getById(id);
         ThrowUtils.throwIf(oldSpaceUser == null, ErrorCode.NOT_FOUND_ERROR);
-        //操作数据库
+        // 操作数据库
         boolean result = spaceUserService.updateById(spaceUser);
         ThrowUtils.throwIf(!result, ErrorCode.OPERATION_ERROR);
         return ResultUtils.success(true);
     }
 
-
     /**
      * 查询我加入的团队空间列表
-     * @param request
-     * @return
      */
     @PostMapping("/list/my")
-    public BaseResponse<List<SpaceUserVO>> listMyTeamSpace(HttpServletRequest request){
+    public BaseResponse<List<SpaceUserVO>> listMyTeamSpace(HttpServletRequest request) {
         User loginUser = userService.getLoginUser(request);
         SpaceUserQueryRequest spaceUserQueryRequest = new SpaceUserQueryRequest();
         spaceUserQueryRequest.setUserId(loginUser.getId());
-        List<SpaceUser> spaceUserList = spaceUserService.list(spaceUserService.getQueryWrapper(spaceUserQueryRequest));
+        List<SpaceUser> spaceUserList = spaceUserService.list(
+                spaceUserService.getQueryWrapper(spaceUserQueryRequest)
+        );
         return ResultUtils.success(spaceUserService.getSpaceUserVOList(spaceUserList));
     }
-
 }
